refactor(query): Split planner/joins.rs into submodules
Extract 3 helper methods from the 2,606-line apply_joins() and distribute
all functions across 7 submodule files:

- mod.rs: apply_joins() main loop + JoinResult + SelectionJoinInfo
- filter_only.rs: apply_filter_only_joins() (extracted block)
- secondary_id.rs: apply_secondary_id_joins() (extracted block)
- aggregate_joins.rs: apply_aggregate_joins() (extracted block)
- filter_relation.rs: apply_filter_relation_join() (moved)
- multi_level.rs: apply_multi_level_sub/filter_joins() (moved)
- mapping.rs: build_scan_mapping_for_join() (moved)
